+++ framework/db-local/src/test/java/is/codion/framework/db/local/DefaultLocalEntityConnectionTest.java
@@ -379,10 +379,10 @@ public class DefaultLocalEntityConnectionTest {
     Select select = Select.where(Condition.custom(Employee.NAME_IS_BLAKE_CONDITION)).build();
     result = connection.select(select);
     Entity emp = result.iterator().next();
-    assertTrue(emp.loaded(Employee.DEPARTMENT_FK));
-    assertTrue(emp.loaded(Employee.MGR_FK));
+    assertNotNull(emp.get(Employee.DEPARTMENT_FK));
+    assertNotNull(emp.get(Employee.MGR_FK));
     emp = emp.referencedEntity(Employee.MGR_FK);
-    assertFalse(emp.loaded(Employee.MGR_FK));
+    assertNull(emp.get(Employee.MGR_FK));
 
     select = Select.where(select.where())
             .fetchDepth(Employee.DEPARTMENT_FK, 0)
@@ -390,8 +390,8 @@ public class DefaultLocalEntityConnectionTest {
     result = connection.select(select);
     assertEquals(1, result.size());
     emp = result.iterator().next();
-    assertFalse(emp.loaded(Employee.DEPARTMENT_FK));
-    assertTrue(emp.loaded(Employee.MGR_FK));
+    assertNull(emp.get(Employee.DEPARTMENT_FK));
+    assertNotNull(emp.get(Employee.MGR_FK));
 
     select = Select.where(select.where())
             .fetchDepth(Employee.DEPARTMENT_FK, 0)
@@ -400,8 +400,8 @@ public class DefaultLocalEntityConnectionTest {
     result = connection.select(select);
     assertEquals(1, result.size());
     emp = result.iterator().next();
-    assertFalse(emp.loaded(Employee.DEPARTMENT_FK));
-    assertFalse(emp.loaded(Employee.MGR_FK));
+    assertNull(emp.get(Employee.DEPARTMENT_FK));
+    assertNull(emp.get(Employee.MGR_FK));
 
     select = Select.where(select.where())
             .fetchDepth(Employee.DEPARTMENT_FK, 0)
@@ -410,10 +410,10 @@ public class DefaultLocalEntityConnectionTest {
     result = connection.select(select);
     assertEquals(1, result.size());
     emp = result.iterator().next();
-    assertFalse(emp.loaded(Employee.DEPARTMENT_FK));
-    assertTrue(emp.loaded(Employee.MGR_FK));
+    assertNull(emp.get(Employee.DEPARTMENT_FK));
+    assertNotNull(emp.get(Employee.MGR_FK));
     emp = emp.referencedEntity(Employee.MGR_FK);
-    assertTrue(emp.loaded(Employee.MGR_FK));
+    assertNotNull(emp.get(Employee.MGR_FK));
 
     select = Select.where(select.where())
             .fetchDepth(Employee.DEPARTMENT_FK, 0)
@@ -422,10 +422,10 @@ public class DefaultLocalEntityConnectionTest {
     result = connection.select(select);
     assertEquals(1, result.size());
     emp = result.iterator().next();
-    assertFalse(emp.loaded(Employee.DEPARTMENT_FK));
-    assertTrue(emp.loaded(Employee.MGR_FK));
+    assertNull(emp.get(Employee.DEPARTMENT_FK));
+    assertNotNull(emp.get(Employee.MGR_FK));
     emp = emp.referencedEntity(Employee.MGR_FK);
-    assertTrue(emp.loaded(Employee.MGR_FK));
+    assertNotNull(emp.get(Employee.MGR_FK));
 
     assertEquals(4, connection.count(where(Employee.ID.in(asList(1, 2, 3, 4)))));
     assertEquals(0, connection.count(where(Employee.DEPARTMENT.isNull())));