+++ changelog.md
@@ -1,19 +1,22 @@
 Codion Change Log
 ==================
 
-## 0.17.19-SNAPSHOT
+## 0.17.19
 ### is.codion.common.db
 - AbstractConnectionPoolWrapper, prevent destroyed connection counter being called twice for each closed connection, see https://github.com/brettwooldridge/HikariCP/issues/2152.
 ### is.codion.framework.domain
 - Condition.customCondition() renamed custom().
 - BlobColumnDefinition removed, ColumnDefinition.lazy() added. Default behaviour for byte array columns being lazy and hidden removed. Related changes.
 - DefaultEntityBuilder, defensive copying of value sets added.
+- Entity.loaded(ForeignKey) removed, only used in tests.
 ### is.codion.framework.db
 - EntityConnection.readBlob() and writeBlob() removed.
 ### is.codion.framework.model
 - EntityTableModel.attributes() added, specifying the attributes to include when selecting. Implemented in SwingEntityTableModel, replacing overridable attributes() method. Related refactoring.
 ### is.codion.swing.framework.model
 - SwingEntityTableModel.getColor() renamed toColor().
+### is.codion.swing.framework.ui
+- EntityPanel.createEditControlPanel() renamed createEditControls().
 
 ## 0.17.18
 - Text.padString() deprecated for removal along with Text.ALIGNMENT, Text.leftPad() and rightPad() added.